requirements.txt
about-time==4.2.1
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
alive-progress==3.1.5
-e git+https://github.com/openvinotoolkit/anomalib.git@5f25510f08dab21bc9419d9c0a64be95728e358e#egg=anomalib
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autograd==1.6.2
av==11.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cma==3.2.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.0
FrEIA==0.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
grapheme==0.6.0
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-resources==6.1.1
ipython==8.19.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonargparse==4.27.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jstyleson==0.0.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.6.9
lazy_loader==0.3
lightning-utilities==0.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
natsort==8.4.0
networkx==2.8.2
ninja==1.10.2.4
nncf==2.7.0
numpy==1.23.4
omegaconf==2.3.0
onnx==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openvino==2023.2.0
openvino-dev==2023.2.0
openvino-telemetry==2023.2.1
packaging==23.2
pandas==1.3.5
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.17.2
pymoo==0.6.1.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
texttable==1.7.0
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.6.13
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchmetrics==0.10.3
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.0
typeshed-client==2.4.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.12
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.9.4
